perf: fix uint boxing in XmlWriter and combine value slice reads during save
Two save-path optimizations:

1. WriteAttribute(string, uint): cast uint to long instead of letting
   XmlWriter.WriteValue(object) box it. XmlWriter has no uint overload,
   so the previous code boxed on every call — ~3.75M allocations for
   the benchmark's cell style IDs.

2. Add ValueSlice.GetCellValueAndShareString that returns both the cell
   value and share-string flag from a single Lut traversal. Use it in
   SheetDataWriter.WriteCellAtPoint to eliminate the second slice read
   that GetShareString required for every value-only cell.

+++ XLibur/Extensions/XmlWriterExtensions.cs
@@ -36,8 +36,10 @@ internal static class XmlWriterExtensions
 
         public void WriteAttribute(string attrName, uint value)
         {
+            // Cast to long to avoid boxing via XmlWriter.WriteValue(object).
+            // XmlWriter has a WriteValue(long) overload but not WriteValue(uint).
             w.WriteStartAttribute(attrName);
-            w.WriteValue(value);
+            w.WriteValue((long)value);
             w.WriteEndAttribute();
         }
